+++ bindings/src/lib.rs
@@ -8,6 +8,7 @@ use crate::models::{PyRegion, PyTokenizedRegion, PyTokenizedRegionSet, PyUnivers
 mod vocab;
 mod tokenizers;
 mod models;
+mod consts;
 
 #[pymodule]
 fn vocab_module(_py: Python, m: &PyModule) -> PyResult<()> {
@@ -26,18 +27,32 @@ fn tokenize_module(_py: Python, m: &PyModule) -> PyResult<()> {
 }
 
 #[pymodule]
+fn consts_module(_py: Python, m: &PyModule) -> PyResult<()> {
+    m.add("PAD_CHR", consts::PAD_CHR)?;
+    m.add("PAD_START", consts::PAD_START)?;
+    m.add("PAD_END", consts::PAD_END)?;
+    m.add("UNKNOWN_CHR", consts::UNKNOWN_CHR)?;
+    m.add("UNKNOWN_START", consts::UNKNOWN_START)?;
+    m.add("UNKNOWN_END", consts::UNKNOWN_END)?;
+    Ok(())
+}
+
+#[pymodule]
 fn genimtools(py: Python, m: &PyModule) -> PyResult<()> {    
     let vocab_module = pyo3::wrap_pymodule!(vocab_module);
     let tokenize_module = pyo3::wrap_pymodule!(tokenize_module);
+    let consts_module = pyo3::wrap_pymodule!(consts_module);
 
     m.add_wrapped(vocab_module)?;
     m.add_wrapped(tokenize_module)?;
+    m.add_wrapped(consts_module)?;
 
     let sys = PyModule::import(py, "sys")?;
     let sys_modules: &PyDict = sys.getattr("modules")?.downcast()?;
 
     sys_modules.set_item("genimtools.vocab", m.getattr("vocab_module")?)?;
     sys_modules.set_item("genimtools.tokenize", m.getattr("tokenize_module")?)?;
+    sys_modules.set_item("genimtools.consts", m.getattr("consts_module")?)?;
 
     Ok(())
 }